+++ __tests__/lib/getNodeVersion.test.ts
@@ -1,7 +1,7 @@
 import semver from 'semver';
 
 import pkg from '../../package.json';
-import getNodeVersion from '../../src/lib/getNodeVersion';
+import { getNodeVersion } from '../../src/lib/getPkgVersion';
 
 describe('#getNodeVersion()', () => {
   it('should extract version that matches range in package.json', () => {
