+++ package.json
@@ -97,7 +97,7 @@
     "build": "tsc",
     "bump": "npm version -m 'build: %s release'",
     "debug": "ts-node src/cli.ts",
-    "lint": "eslint . bin/rdme bin/set-version-output --ext .js,.ts",
+    "lint": "eslint . bin/rdme --ext .js,.ts",
     "lint-docs": "npx alex . && npm run prettier:docs",
     "postversion": "git tag $npm_package_version",
     "prebuild": "rm -rf dist/",
